col2secをテキストのみコンポーネント化しました。
