+++ src/components/section/Col2.module.css
@@ -1,0 +1,15 @@
+.col2{
+    display: grid;
+    grid-template-columns: repeat(2,1fr);
+    align-items: center;
+    gap:var(--gap-48px);
+    img{
+      width:100%;
+      height:auto;
+    }
+    &.dark {
+      background: var(--bg-color-200);
+      box-shadow: 0 0 0 100vmax var(--bg-color-200);
+      clip-path: inset(0 -100vmax);
+    }
+  }
