+++ src/data/topCafeData.tsx
@@ -1,0 +1,10 @@
+const content = {
+  title: "カフェスペース",
+  subTitle: "焼きたてパンとコーヒーを楽しむ時間",
+  text : [
+    "店内には木の温もりを感じるイートインスペースをご用意。 焼きたてのパンを、挽きたての香り高いコーヒーとともにお楽しみいただけます。 日常を忘れて、ゆっくりとした時間をお過ごしください。"
+  ],
+  href : "/cafe/",
+  btnText : "カフェについて詳しく"
+}
+export default content;
